ZERO Particles -  thanks to the creator of poeNullParticles!
- added an option to remove all particles from the game with just one
click, (i have used poeNullParticles to set all files to 0, a lot of
them...:D ).

There are two separated folders, one with original game files
"restoreParticles" and that should not be touched, and another one with
all files set to 0 value .."replaceParticles" .
The good part of this is that if you want to have your skill untouched,
just go to "replaceParticles"..and delete the folder that contains your
skill/effect you want to see in game...or any other effects.
All other options are still available ofc, and i already removed them
from "replaceParticles", to have them separately.

enjoy,